+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Zichtbare craftables verhoogd in specifieke benches
 - Weedplanten voeding gaat trager naar beneden
 - Vuilbak doorzoek taskbar gaat sneller
+- Voertuig status decrease een beetje meer gebalanceerd
 
 ## Fixed
 - Fix speedbans wanneer je uit wagen vliegt of onder de grond valt